# Simplify or hide lots of log output.