@isaacabraham 2020-11-22
There will probably be some inconsistencies on this but I think for the main use case it'll help.