@forki 2020-11-22
Please keep those coming.

Isaac Abraham <notifications@github.com> schrieb am So., 22. Nov. 2020,
19:01:

> There will probably be some inconsistencies on this but I think for the
> main use case it'll help.
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/fsprojects/Paket/pull/3952#issuecomment-731785509>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAOANEXSGKOBA4H6HG77QTSRFGX5ANCNFSM4T6AODMA>
> .
>